Fix options => init in tests; refactor prepare tests
