+++ test/prepare/helpers.js
@@ -1,7 +1,7 @@
 var chai = require('chai');
 var config = require('../config');
 var expect = chai.expect;
-var options = require('../../dist/options');
+var init = require('../../dist/init');
 var prepareHelpers = require('../../dist/prepare/helpers');
 var sinon = require('sinon');
 var sinonChai = require('sinon-chai');
@@ -10,17 +10,18 @@ chai.use(sinonChai);
 describe ('prepare/helpers', () => {
   var opts;
   before(() => {
-    opts = options(config.fixtureOpts);
+    opts = init(config.fixtureOpts);
+    return opts;
   });
   it ('should register handlebars-layouts helpers', () => {
-    return prepareHelpers(opts).then(() => {
-      expect(opts.handlebars.helpers).to.contain.keys(
+    return opts.then(prepareHelpers).then(options => {
+      expect(options.handlebars.helpers).to.contain.keys(
         'embed', 'block', 'content');
     });
   });
   it ('should register pattern helpers', () => {
-    return prepareHelpers(opts).then(() => {
-      expect(opts.handlebars.helpers).to.contain.keys(
+    return opts.then(prepareHelpers).then(options => {
+      expect(options.handlebars.helpers).to.contain.keys(
         'pattern', 'patternSource');
     });
   });
@@ -28,14 +29,14 @@ describe ('prepare/helpers', () => {
     opts.helpers = {
       foo: function () { return 'foo'; }
     };
-    return prepareHelpers(opts).then(() => {
-      expect(opts.handlebars.helpers).to.contain.keys('foo');
+    return init(opts).then(prepareHelpers).then(options => {
+      expect(options.handlebars.helpers).to.contain.keys('foo');
     });
   });
   it ('should register helper files', () => {
     opts.helpers = config.fixturePath('helpers/**/*.js');
-    return prepareHelpers(opts).then(() => {
-      expect(opts.handlebars.helpers).to.contain.keys('toFraction',
+    return init(opts).then(prepareHelpers).then(options => {
+      expect(options.handlebars.helpers).to.contain.keys('toFraction',
         'toJSON',
         'toSlug',
         'random',
@@ -44,22 +45,7 @@ describe ('prepare/helpers', () => {
     });
   });
   describe ('namespace conflicts', () => {
-    var opts, logStub;
-    before (() => {
-      opts = Object.assign({}, options(config.fixtureOpts));
-      logStub = sinon.stub(opts.debug, 'logFn');
-    });
-    it ('should report namespace conflicts', () => {
-      opts.handlebars.registerPartial('pattern', 'empty');
-
-      return prepareHelpers(opts).then(() => {
-        logStub.restore();
-        opts.handlebars.unregisterPartial('pattern');
-        expect(logStub).to.have.been.calledOnce;
-      });
-    });
     it ('should have this test expanded when error-handling done');
-
   });
 
 });
