+++ test/prepare/partials.js
@@ -1,29 +1,32 @@
-/* global describe, it */
 var chai = require('chai');
 var config = require('../config');
 var expect = chai.expect;
-var options = require('../../dist/options');
+var init = require('../../dist/init');
 var preparePartials = require('../../dist/prepare/partials');
 
 describe ('prepare/partials', () => {
-  const opts = options(config.fixtureOpts);
+  var opts;
+  before(() => {
+    opts = init(config.fixtureOpts);
+    return opts;
+  });
   it ('should register layout partials', () => {
-    return preparePartials(opts). then(() => {
-      expect(opts.handlebars.partials).to.contain.keys(
+    return opts.then(preparePartials).then(options => {
+      expect(options.handlebars.partials).to.contain.keys(
         'collection', 'default', 'page');
     });
   });
   it ('should register partial partials', () => {
-    return preparePartials(opts). then(() => {
-      expect(opts.handlebars.partials).to.contain.keys(
+    return opts.then(preparePartials). then(options => {
+      expect(options.handlebars.partials).to.contain.keys(
         'header', 'menu', 'nested.thing');
     });
   });
   it ('should register patterns files as partials', () => {
-    return preparePartials(opts). then(() => {
-      expect(opts.handlebars.partials).to.contain.keys(
+    return opts.then(preparePartials). then(options => {
+      expect(options.handlebars.partials).to.contain.keys(
         'patterns.components.orange', 'patterns.components.button.aardvark');
-      expect(opts.handlebars.partials).to.contain.keys(
+      expect(options.handlebars.partials).to.contain.keys(
         'patterns.pink',
         'patterns.components.button.base',
         'patterns.components.button.color-variation'
